+++ NEWS.md
@@ -1,3 +1,34 @@
+# Version 0.16.0 [2025-05-18]
+
+## New Features
+
+ * Add `handler_slowdown()` for slowing down a full progress run a
+   certain amount of time (in seconds).
+
+## Workarounds
+
+ * RStudio v2025.05.0 introduced new default handlers for messages and
+   warnings for the RStudio Console, which came with a bug causing
+   console output to stop working in some cases. This bug was fixed in
+   RStudio v2025.05.1. If you run v2025.05.0, **progressr** implements
+   a workaround.
+
+## Documentation
+
+ * Replace the long, single introductory vignette with eight separate
+   vignettes.
+
+## Bug Fixes
+
+ * Package failed to load in the Positron environment due to an
+   "object 'RStudio.Version' of mode 'function' was not found" error.
+
+ * In Positron, messages produced while progress is reported were
+   prepended by a newline to standard output, even when there is no
+   stdout buffered. This is due to a bug in Positron, which this version
+   of the package works around.
+
+
 # Version 0.15.1 [2024-11-20]
 
 ## Bug Fixes